Added navigation view and pages
