+++ UStealth.WinUI/App.xaml.cs
@@ -27,7 +27,7 @@ namespace UStealth.WinUI
     public partial class App : Application
     {
         private Window? _window;
-
+        public Frame RootFrame { get; set; }
         /// <summary>
         /// Initializes the singleton application object.  This is the first line of authored code
         /// executed, and as such is the logical equivalent of main() or WinMain().
