+++ UStealth.WinUI/MainWindow.xaml
@@ -23,87 +23,22 @@
             <TitleBar.IconSource>
                 <ImageIconSource x:Name="AppIcon" />
             </TitleBar.IconSource>
-            <StackPanel
-                VerticalAlignment="Center"
-                Orientation="Horizontal"
-                Spacing="8">
-                <Button
-                    x:Name="RefreshButton"
-                    Width="100"
-                    Height="32"
-                    HorizontalAlignment="Right"
-                    Click="RefreshButton_Click"
-                    Content="Refresh" />
-            </StackPanel>
         </TitleBar>
 
-        <Grid Grid.Row="1" Padding="16">
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" />
-                <RowDefinition Height="*" />
-            </Grid.RowDefinitions>
-            <ProgressRing x:Name="ProgressRing" IsActive="False" />
-            <tableView:TableView
-                x:Name="DrivesTableView"
-                Grid.Row="1"
-                Margin="0,16,0,0"
-                x:DataType="local:DriveInfoModel"
-                AutoGenerateColumns="False"
-                ItemsSource="{x:Bind Drives}"
-                SelectionMode="Single">
-                <tableView:TableView.Columns>
-                    <tableView:TableViewCheckBoxColumn
-                        Binding="{Binding IsSystemDrive}"
-                        Header="System Drive"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Binding="{Binding DriveLetter}"
-                        Header="Drive Letter"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Width="80"
-                        Binding="{Binding Interface}"
-                        Header="Interface"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Binding="{Binding VolumeLabel}"
-                        Header="Volume Label"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Binding="{Binding Format}"
-                        Header="Format"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Width="200"
-                        Binding="{Binding Model}"
-                        Header="Model"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Binding="{Binding MediaType}"
-                        Header="Media type"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTextColumn
-                        Binding="{Binding Size}"
-                        Header="Size"
-                        IsReadOnly="True" />
-                    <tableView:TableViewTemplateColumn x:Name="StatusTemplate" Header="Status">
-                        <tableView:TableViewTemplateColumn.CellTemplate>
-                            <DataTemplate>
-                                <StackPanel>
-                                    <ToggleSwitch
-                                        x:Name="StatusToggle"
-                                        IsOn="{Binding Status, Mode=TwoWay, Converter={StaticResource StatusToIsOnConverter}}"
-                                        OffContent="NORMAL"
-                                        OnContent="HIDDEN"
-                                        Toggled="StatusToggle_OnToggled"
-                                        Visibility="{Binding Status, Converter={StaticResource StatusToTemplateColumnVisibilityConverter}}" />
-                                    <TextBlock Text="{Binding Status}" Visibility="{Binding Status, Converter={StaticResource UnknownStatusVisibilityConverter}}" />
-                                </StackPanel>
-                            </DataTemplate>
-                        </tableView:TableViewTemplateColumn.CellTemplate>
-                    </tableView:TableViewTemplateColumn>
-                </tableView:TableView.Columns>
-            </tableView:TableView>
-        </Grid>
+        <NavigationView
+            Grid.Row="1"
+            IsBackButtonVisible="Collapsed"
+            PaneDisplayMode="Top"
+            SelectionChanged="NavigationView_OnSelectionChanged"
+            SelectionFollowsFocus="Enabled">
+            <NavigationView.MenuItems>
+                <NavigationViewItem x:Name="Main" Tag="Main">
+                    <NavigationViewItem.Icon>
+                        <FontIcon Glyph="&#xEC4E;" />
+                    </NavigationViewItem.Icon>
+                </NavigationViewItem>
+            </NavigationView.MenuItems>
+            <Frame x:Name="rootFrame" />
+        </NavigationView>
     </Grid>
 </Window>
