+++ UStealth.WinUI/Pages/SettingsPage.xaml.cs
@@ -1,0 +1,18 @@
+using Microsoft.UI.Xaml.Controls;
+
+// To learn more about WinUI, the WinUI project structure,
+// and more about our project templates, see: http://aka.ms/winui-project-info.
+
+namespace UStealth.WinUI.Pages
+{
+    /// <summary>
+    /// An empty page that can be used on its own or navigated to within a Frame.
+    /// </summary>
+    public sealed partial class SettingsPage : Page
+    {
+        public SettingsPage()
+        {
+            InitializeComponent();
+        }
+    }
+}
