+++ UStealth.WinUI/UStealth.WinUI.csproj
@@ -13,6 +13,10 @@
     <Nullable>enable</Nullable>
     <WindowsAppSDKSelfContained>true</WindowsAppSDKSelfContained>
   </PropertyGroup>
+  <ItemGroup>
+    <None Remove="Pages\MainPage.xaml" />
+    <None Remove="SettingsPage.xaml" />
+  </ItemGroup>
 
   <ItemGroup>
     <Content Include="Assets\SplashScreen.scale-200.png" />
@@ -43,6 +47,16 @@
   <ItemGroup>
     <ProjectReference Include="..\UStealth.DriveHelper\UStealth.DriveHelper.csproj" />
     <ProjectReference Include="..\WinUI.TableView\WinUI.TableView.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <Page Update="SettingsPage.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+  </ItemGroup>
+  <ItemGroup>
+    <Page Update="Pages\MainPage.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
   </ItemGroup>
 
   <!--
